Support for VCS backed workspaces
Add the option to create VCS backed workspaces using k8s. This can be
achieved by adding a `vcs` section in the workspace CR spec like this:
```
 vcs:
   token_id: "oauth client id taken from TFC (or TFE)"
   repo_identifier: "repository identifier (i.e koikonom/terraform-nullresource-example")
```

The operator now loops every 60 seconds and if it detects a run
that took place "out of band" (i.e a user queued it in the UI,
or new code was pushed to the repository), then the operator will
update the outputs in the k8s resource accordingly.