Store workspace outputs as secrets (#80)
* Store workspace outputs as secrets

Currently the operator stores outputs as configmaps. This can be
undesirable for some environments since output values can be considered
sensitive. This change creates k8s secrets instead of ConfigMaps to
satisfy this usecase.

During review of #80 it was discovered that v1beta1 of the CRD schema
was not compatible with setting default values, like we do for the
workspace's default version. Since setting the CRD schema version to v1
instead would result in dropping support for k8s versions before 1.16.0 
we'll allow the schema to support empty terraform versions and handle 
the lack of a value in the code.