Revert to parsing outputs via the statefile
Although it's easier to retrieve state outputs using the TFC API, there
are some limitations in the way the results are parsed that causes error
if the output is not a string.

The old way of using terraform's code to parse the statefile is more
versatile and since version 0.14 is both backward and forward compatible
we no longer have to update the operator for each new terraform release.

This change replaces the output retrieving logic with the old one, but
retains the ability to set a terraform version for a workspace.